+++ pysatCDAAC/instruments/__init__.py
@@ -5,7 +5,7 @@ Each instrument is contained within a subpackage of this set.
 
 """
 
-__all__ = ['cosmic_gps']
+__all__ = ['cosmic_gps', 'cosmic2_ivm']
 
 for inst in __all__:
     exec("from pysatCDAAC.instruments import {x}".format(x=inst))